feat(error-handling): add catchError func to static handler
