+++ src/modules/try-catch/index.ts
@@ -2,3 +2,4 @@ export * from './try-catch-emitter';
 export * from './try-catch-exception.interface';
 export * from './try-catch-options.interface';
 export * from './try-catch.decorator';
+export * from './catch-error.util';
